story-refactor(dm): add stdint and trace policy to C++17 examples

+++ examples/cpp/hsm_echo.cpp
@@ -127,9 +127,9 @@ private:
     static void conn_thunk(struct ev_loop* loop, ev_io* w, int revents) noexcept;
     static void* client_main(void* arg) noexcept;
 
-    void accept_cb(ev_io& w, int revents) noexcept;
-    void timeout_cb(ev_timer& w, int revents) noexcept;
-    void handle_conn(Conn& c, ev_io& w, int revents) noexcept;
+    void accept_cb(ev_io& w, uint32_t revents) noexcept;
+    void timeout_cb(ev_timer& w, uint32_t revents) noexcept;
+    void handle_conn(Conn& c, ev_io& w, uint32_t revents) noexcept;
 
     uint16_t alloc_conn() noexcept;
     void teardown(Conn& c) noexcept;
@@ -276,13 +276,13 @@ void HsmEchoServer::conn_thunk(struct ev_loop*, ev_io* w, int revents) noexcept
     {
         if (&c.io == w)
         {
-            self->handle_conn(c, *w, revents);
+            self->handle_conn(c, *w, static_cast<uint32_t>(revents));
             return;
         }
     }
 }
 
-void HsmEchoServer::handle_conn(Conn& c, ev_io& w, int revents) noexcept
+void HsmEchoServer::handle_conn(Conn& c, ev_io& w, uint32_t revents) noexcept
 {
     if (revents & EV_ERROR)
     {
@@ -340,7 +340,7 @@ void HsmEchoServer::handle_conn(Conn& c, ev_io& w, int revents) noexcept
     }
 }
 
-void HsmEchoServer::accept_cb(ev_io& w, int revents) noexcept
+void HsmEchoServer::accept_cb(ev_io& w, uint32_t revents) noexcept
 {
     (void)revents;
 
@@ -382,7 +382,7 @@ void HsmEchoServer::accept_cb(ev_io& w, int revents) noexcept
     }
 }
 
-void HsmEchoServer::timeout_cb(ev_timer&, int) noexcept
+void HsmEchoServer::timeout_cb(ev_timer&, uint32_t) noexcept
 {
     loop_.break_loop();
 }
